update tools/checkdomains.sh update tmp log routines

+++ tools/checkdomains.sh
@@ -15,6 +15,7 @@ WHOIS_SHOWREGISTRAR='y'
 WHOIS_NAMESERVER='8.8.8.8'
 
 CHECKDOMAINS_DEBUG='n'
+DELETE_TMPLOGS='y'
 CTMPDIR=/home/checkdomainstmp
 ##########################################################################
 if [ -f "/etc/centminmod/custom_config.inc" ]; then
@@ -43,7 +44,7 @@ if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
 LISTDOMAINS=$(egrep -rn 'http:|https:' /usr/local/src/centminmod/ | egrep -v 'http://ftp.osuosl.org|\${HTUSER}|\$request_uri|\$vhostname|\${vhostname}|rpm.axivo.com|foo.bar|master.ourdelta.org|newdomain1.com|apt.sw.be|medium.com|href=|my.incapsula.com|#|echo|cecho|<li>|<li class|centos.alt.ru|<|>|\(|\[|\)|\]|<html|<!DOCTYPE|nginx.org|centminmod.com|centmin.com|centmin.sh|github.com|php.net|yum.mariadb.org|apache.mirror.uber.com.au' | sed -e "s|<||g" -e "s|'||g" -e "s|\| bash -s stable||g" | grep -Eo '(http|https|ftp)://[^/"]+' | sed -e "s|http:\/\/||g" -e "s|https:\/\/||g" | sort | uniq -c | sort -rn | awk '{print $2}')
 fi
 
-OTHERDOMAINS='nginx.org centminmod.com centmin.com centmin.sh github.com php.net yum.mariadb.org apache.mirror.uber.com.au'
+OTHERDOMAINS='nginx.org centminmod.com centmin.com centmin.sh github.com php.net yum.mariadb.org'
 
 for d in ${OTHERDOMAINS[@]}; do
   echo "----------"
@@ -84,7 +85,7 @@ for d in ${OTHERDOMAINS[@]}; do
         done
       done
     fi
-    if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
+    if [[ "$DELETE_TMPLOGS" = [yY] ]]; then
       rm -rf "${CTMPDIR}/${d}.txt"
     fi
   elif [[ "$WHOISBIN" = 'whois' && "$WHOISOPT" = ' -n' ]]; then
@@ -124,7 +125,9 @@ for d in ${OTHERDOMAINS[@]}; do
         echo -n "$d "
         echo "$ipaddr $country $org" | tr '\n\r' ' ' | tr -s ' '
         echo
-        rm -rf "${CTMPDIR}/${d}-ip.txt"
+        if [[ "$DELETE_TMPLOGS" = [yY] ]]; then        
+          rm -rf "${CTMPDIR}/${d}-ip.txt"
+        fi
       fi
     done
     if [[ "$WHOIS_SHOWNS" = [yY] ]]; then
@@ -137,7 +140,7 @@ for d in ${OTHERDOMAINS[@]}; do
         done
       done
     fi
-    if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
+    if [[ "$DELETE_TMPLOGS" = [yY] ]]; then
       rm -rf "${CTMPDIR}/${d}.txt"
     fi
   elif [[ "$WHOISBIN" = 'whois' ]]; then
@@ -177,7 +180,7 @@ for d in ${OTHERDOMAINS[@]}; do
         done
       done
     fi
-    if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
+    if [[ "$DELETE_TMPLOGS" = [yY] ]]; then
       rm -rf "${CTMPDIR}/${d}.txt"
     fi
   fi
@@ -224,7 +227,7 @@ if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
           done
         done
       fi
-      if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
+      if [[ "$DELETE_TMPLOGS" = [yY] ]]; then
         rm -rf "${CTMPDIR}/${d}.txt"
       fi
     elif [[ "$WHOISBIN" = 'whois' && "$WHOISOPT" = ' -n' ]]; then
@@ -264,7 +267,9 @@ if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
           echo -n "$d "
           echo "$ipaddr $country $org" | tr '\n\r' ' ' | tr -s ' '
           echo
-          rm -rf "${CTMPDIR}/${d}-ip.txt"
+          if [[ "$DELETE_TMPLOGS" = [yY] ]]; then
+            rm -rf "${CTMPDIR}/${d}-ip.txt"
+          fi
         fi
       done
       if [[ "$WHOIS_SHOWNS" = [yY] ]]; then
@@ -277,7 +282,7 @@ if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
           done
         done
       fi
-      if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
+      if [[ "$DELETE_TMPLOGS" = [yY] ]]; then
         rm -rf "${CTMPDIR}/${d}.txt"
       fi
     elif [[ "$WHOISBIN" = 'whois' ]]; then
@@ -317,7 +322,7 @@ if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
           done
         done
       fi
-      if [[ "$CHECKDOMAINS_DEBUG" != [yY] ]]; then
+      if [[ "$DELETE_TMPLOGS" = [yY] ]]; then
         rm -rf "${CTMPDIR}/${d}.txt"
       fi
     fi
